frser: support for serprog-add-on of opbuf: poll (+w dly)
This is the support to complement the poll support i've released
https://github.com/urjaman/flashrom/commits/serprog-improvements
in that branch. After seeing what has been said on the
mailing list on other speed related changes ...
(Whilst having "Speed." in bold at flashrom.org )
I'm not going to bother with upstreaming anything at
all from that branch. Just too much of a pain.
Use it if you want LPC/FWH or any decency of speed.
Or explosions, it could always be explosions...