flash: change default port state from tri-state to pullup
This fixes the "random" brown-out-resets i encountered while
probing for SPI chips (even the plausible bustypes test)
with a 5V parallel chip attached. Propably the tri-state
wasnt the best for the parallel chip and it thought it
was enabled or something and pins fought and the USB power
supply isnt the best (and AVR pin drive strenght is good when 5V)
so BOR.

I checked the other drivers that they didnt assume anything
(badly) about the PORT state being zero. Seems to be all ok.